move tests into a single place
