+++ package.json
@@ -2,15 +2,15 @@
     "name": "root",
     "private": "true",
     "scripts": {
-        "check"  : "pnpm run all check",
-        "test"   : "pnpm run all test",
-        "all"    : "pnpm run --workspace-concurrency 1 --bail false --filter \"./packages/*\"",
-        "changed": "pnpm run --workspace-concurrency 1 --bail false --filter \"./packages/*[origin/main]\""
+        "check": "pnpm exec tsc",
+        "test" : "pnpm exec vitest --run --threads false"
     },
     "workspaces": [ "packages/*" ],
     "packageManager": "pnpm@8.6.12",
     "devDependencies": {
         "astro": "3",
+        "vitest": "0.34",
+        "typescript": "5.2",
         "@changesets/cli": "^2.26.2",
         "@changesets/changelog-github": "^0.4.8"
     }
