+++ packages/global/package.json
@@ -27,8 +27,7 @@
     }
   },
   "scripts": {
-    "check": "tsc",
-    "test": "vitest --run --threads false"
+    "test": "pnpm -w test tests/global.test.ts"
   },
   "type": "module",
   "peerDependencies": {
@@ -37,8 +36,6 @@
   "devDependencies": {
     "@types/node": "20",
     "astro": "3",
-    "astro-global": "workspace:*",
-    "typescript": "5.2",
-    "vitest": "0.34"
+    "astro-global": "workspace:*"
   }
 }
