+++ packages/prerender-patterns/package.json
@@ -21,14 +21,11 @@
     ".": "./integration.ts"
   },
   "scripts": {
-    "check": "tsc",
-    "test": "vitest --run --threads false"
+    "test": "pnpm -w test tests/prerender-patterns.test.ts"
   },
   "type": "module",
   "devDependencies": {
     "@types/node": "20",
-    "astro": "3",
-    "typescript": "5.2",
-    "vitest": "0.34"
+    "astro": "3"
   }
 }
