+++ tests/fixtures/adds-to-head/astro.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "astro/config"
 import mdx from "@astrojs/mdx"
-import addsToHead from "../../../integration.ts"
+import addsToHead from "astro-adds-to-head"
 
 export default defineConfig({
     integrations: [ mdx(), addsToHead() ]
