+++ tests/fixtures/adds-to-head/package.json
@@ -3,6 +3,7 @@
   "private": true,
   "dependencies": {
     "astro": "3",
+    "astro-adds-to-head": "workspace:*",
     "@astrojs/mdx": "1"
   }
 }
