+++ tests/fixtures/adds-to-head/src/components/HasScript.astro
