+++ tests/fixtures/adds-to-head/src/components/HasStyle.astro
