+++ tests/fixtures/adds-to-head/src/components/IndirectRenderer.astro
