+++ tests/fixtures/adds-to-head/src/content/blog/promo/launch-week-component.mdx
