+++ tests/fixtures/adds-to-head/src/pages/index.astro
