+++ tests/fixtures/global/astro.config.ts
