+++ tests/fixtures/global/src/components/Counter.jsx
