+++ tests/fixtures/global/src/components/ServerRenderOnly.jsx
