+++ tests/fixtures/global/src/pages/[rest].astro
