+++ tests/fixtures/global/src/pages/mdx-page.mdx
