+++ tests/fixtures/global/tsconfig.json
