+++ tests/fixtures/prerender-patterns/package.json
@@ -2,6 +2,7 @@
   "name": "@test/prerender-patterns",
   "private": true,
   "dependencies": {
-    "astro": "3"
+    "astro": "3",
+    "astro-prerender-patterns": "workspace:*"
   }
 }
