+++ tests/fixtures/prerender-patterns/source/env.d.ts
