+++ tests/fixtures/prerender-patterns/src/not-pages/added-by-integration.astro
