+++ tests/fixtures/prerender-patterns/src/pages/404.astro
