+++ tests/fixtures/prerender-patterns/src/pages/500.astro
