+++ tests/fixtures/prerender-patterns/src/pages/endpoint-default.ts
