+++ tests/fixtures/prerender-patterns/src/pages/endpoint-set-to-prerender.ts
