+++ tests/fixtures/prerender-patterns/src/pages/endpoint-set-to-render-on-demand.ts
