+++ tests/fixtures/prerender-patterns/src/pages/page-default.astro
