+++ tests/fixtures/prerender-patterns/src/pages/page-set-to-prerender.astro
