+++ tests/fixtures/prerender-patterns/src/pages/page-set-to-render-on-demand.astro
