+++ tests/fixtures/prerender-patterns/src/pages/page-with-script.astro
