+++ tests/fixtures/scope/astro.config.ts
