+++ tests/fixtures/scope/package.json
