+++ tests/fixtures/scope/src/pages/page-a.astro
