+++ tests/fixtures/scope/src/pages/page-b.astro
